pyproject.toml
[project]
name = "dcie"
version = "0.1.0"
description = "Document Compliance Identification Engine/dcie"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "asyncpg>=0.30.0",
    "chardet>=5.2.0",
    "fastapi>=0.116.1",
    "funasr>=1.2.7",
    "loguru>=0.7.3",
    "nanoid>=2.0.0",
    "paddleocr>=3.2.0",
    "pyahocorasick>=2.2.0",
    "pydantic-settings>=2.10.1",
    "pytesseract>=0.3.13",
    "python-docx>=1.2.0",
    "python-multipart>=0.0.20",
    "python-pptx>=1.0.2",
    "redis>=6.4.0",
    "sqlalchemy>=2.0.43",
    "tomli>=2.2.1",
    "transformers>=4.56.1",
    "uvicorn>=0.35.0",
]
